+++ src/main/java/com/bbangle/bbangle/config/security/WebMvcConfig.java
@@ -1,0 +1,39 @@
+package com.bbangle.bbangle.config.security;
+
+import static org.springframework.http.HttpMethod.DELETE;
+import static org.springframework.http.HttpMethod.GET;
+import static org.springframework.http.HttpMethod.PATCH;
+import static org.springframework.http.HttpMethod.POST;
+import static org.springframework.http.HttpMethod.PUT;
+
+import com.bbangle.bbangle.auth.oauth.OauthServerTypeConverter;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.format.FormatterRegistry;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    private static final String[] ALLOWED_ORIGINS = new String[]{
+        "http://localhost:3000",
+        "https://www.bbanggree.com",
+        "https://api.bbanggree.com",
+        "https://develop.bbanggree.com"
+    };
+
+    @Override
+    public void addFormatters(FormatterRegistry registry) {
+        registry.addConverter(new OauthServerTypeConverter());
+    }
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+            .allowedOrigins(ALLOWED_ORIGINS)
+            .allowedMethods(GET.name(), POST.name(), PUT.name(), DELETE.name(), PATCH.name())
+            .allowCredentials(true)
+            .exposedHeaders("*");
+    }
+
+}
